+ Bbox is now as class + redesigned Trimesh collision and convexhull shape + trimesh collision shape now with LIMIT 32767 indices

+++ trunk/AGFX_jF3D_GraphicsEngine/src/AGFX/F3D/Mesh/TF3D_Mesh.java
@@ -120,7 +120,7 @@ public class TF3D_Mesh
 	
 		String mat_name = "none";
 
-		this.data.SetFacesCount(PARSER.GetBlocksCount());
+		this.data.SetFacesCount((short)PARSER.GetBlocksCount());
 
 		for (int i = 0; i < PARSER.GetBlocksCount(); i++)
 		{
